requirements.txt
Flask>=2.0
requests
beautifulsoup4
dateparser
APScheduler
playwright
cryptography
pytest
google-genai
pypdf
python-docx

